+++ selectrum.el
@@ -416,6 +416,15 @@ This is used to implement `selectrum-repeat'.")
 (defvar selectrum--active-p nil
   "Non-nil means we are in a Selectrum session currently.")
 
+(defvar selectrum--minibuffer nil
+  "Minibuffer currently in use.")
+
+(defvar selectrum--current-candidate-bounds nil
+  "Cons cell of start and end of current candidate in minibuffer.")
+
+(defvar selectrum--ensure-centered-timer nil
+  "Timer to run `selectrum--ensure-current-candidate-centered'.")
+
 ;;;; Hook functions
 
 (defun selectrum--count-info ()
@@ -426,6 +435,55 @@ This is used to implement `selectrum-repeat'.")
       ('matches         (format "%-4d " total))
       ('current/matches (format "%-6s " (format "%d/%d" current total)))
       (_                ""))))
+
+(defun selectrum--ensure-current-candidate-centered ()
+  "\"Scroll\" the minibuffer if needed to ensure current candidate visible.
+This needs to be run on an idle timer since `posn-at-point' won't
+work before redisplay.
+
+Note that the current implementation produces an undesirable
+gradual scrolling effect when browsing long candidates. This is
+hard to avoid since Emacs does not have a good interface for
+figuring out how many lines text will actually take up without
+just rendering it to the screen and then checking."
+  (when (and selectrum--active-p
+             (car selectrum--current-candidate-bounds)
+             (cdr selectrum--current-candidate-bounds))
+    (with-current-buffer selectrum--minibuffer
+      (let ((inhibit-read-only t))
+        (when (let ((start
+                     (cdr
+                      (posn-actual-col-row
+                       (posn-at-point
+                        (car selectrum--current-candidate-bounds)))))
+                    (end
+                     (cdr
+                      (posn-actual-col-row
+                       (posn-at-point
+                        (cdr selectrum--current-candidate-bounds)))))
+                    (showing-everything-p
+                     (posn-at-point (point-max)))
+                    (height (1- (window-body-height))))
+                (and (not showing-everything-p)
+                     (or (not start)
+                         (> start (/ height 2))
+                         (and (or (not end)
+                                  (>= end height))
+                              (> start 1)))))
+          (save-excursion
+            (goto-char (1+ selectrum--end-of-input-marker))
+            (delete-region
+             (point)
+             (save-excursion
+               (end-of-visual-line)
+               (condition-case _
+                   (forward-char)
+                 (end-of-buffer))
+               (point))))
+          ;; Redisplay and do it again.
+          (setq selectrum--ensure-centered-timer
+                (run-with-idle-timer
+                 0 nil #'selectrum--ensure-current-candidate-centered)))))))
 
 (defun selectrum--minibuffer-post-command-hook ()
   "Update minibuffer in response to user input."
@@ -521,6 +579,7 @@ This is used to implement `selectrum-repeat'.")
              (minibuffer-prompt-end) bound
              '(face selectrum-current-candidate)))
           (let ((index 0))
+            (setq selectrum--current-candidate-bounds (cons nil nil))
             (dolist (candidate (funcall
                                 selectrum-highlight-candidates-function
                                 input
@@ -552,6 +611,9 @@ This is used to implement `selectrum-repeat'.")
                    'selectrum-current-candidate
                    'append displayed-candidate))
                 (insert "\n")
+                (when (equal index highlighted-index)
+                  (setf (car selectrum--current-candidate-bounds)
+                        (point-marker)))
                 (when selectrum-show-indices
                   (let* ((abs-index (+ index first-index-displayed))
                          (num (number-to-string (1+ abs-index)))
@@ -567,6 +629,9 @@ This is used to implement `selectrum-repeat'.")
                       'face
                       'minibuffer-prompt))))
                 (insert displayed-candidate)
+                (when (equal index highlighted-index)
+                  (setf (cdr selectrum--current-candidate-bounds)
+                        (point-marker)))
                 (when right-margin
                   (let ((ol (make-overlay (point) (point))))
                     (overlay-put
@@ -586,7 +651,13 @@ This is used to implement `selectrum-repeat'.")
         (set-marker-insertion-type selectrum--end-of-input-marker t)
         (selectrum--fix-set-minibuffer-message))
       (when keep-mark-active
-        (setq deactivate-mark nil)))))
+        (setq deactivate-mark nil))
+      (when selectrum--ensure-centered-timer
+        (cancel-timer selectrum--ensure-centered-timer)
+        (setq selectrum--ensure-centered-timer nil))
+      (setq selectrum--ensure-centered-timer
+            (run-with-idle-timer
+             0 nil #'selectrum--ensure-current-candidate-centered)))))
 
 (defun selectrum--minibuffer-exit-hook ()
   "Clean up Selectrum from the minibuffer, and self-destruct this hook."
@@ -610,6 +681,7 @@ non-nil, means the user has to select one of the candidates
 provided, rather than providing one of their own."
   (add-hook
    'minibuffer-exit-hook #'selectrum--minibuffer-exit-hook nil 'local)
+  (setq selectrum--minibuffer (current-buffer))
   (setq selectrum--match-required-p require-match)
   (setq selectrum--start-of-input-marker (point-marker))
   (if selectrum--repeat